PUT and DELETE methods updation for user

+++ src/URLShortener.API/Application/Interfaces/IUserService.cs
@@ -5,8 +5,8 @@ namespace URLShortener.API.Application.Interfaces;
 public interface IUserService
 {
     Task<IEnumerable<UserDto>> GetAllAsync();
-    Task<UserDto?> GetByIdAsync(short id);
+    Task<UserDto?> GetByIdAsync(int id);
     Task<UserDto> CreateAsync(UserRegisterDto userRegisterDto);
-    Task<bool> UpdateAsync(short id, UserUpdateDto userUpdateDto);
-    Task<bool> DeleteAsync(short id);
+    Task<bool> UpdateAsync(int id, UserUpdateDto userUpdateDto);
+    Task<bool> DeleteAsync(int id);
 }
